requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
billiard==4.2.1
channels==4.2.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
django==4.2.23
gunicorn==23.0.0
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
